change constant to int

+++ src/main/java/frc/robot/constants/RollerConstants.java
@@ -6,7 +6,7 @@ package frc.robot.constants;
 
 /** Add your docs here. */
 public class RollerConstants {
-    public static final double kRollerMotor = 12;
+    public static final int kRollerMotor = 12;
 
     public static final double kRollerSpeed = 0.5;
 }
